Remove output nodes for now; show as view options for current node
This makes more sense for the current outputs, which are mostly visual.
They'll now act as representation viewers for the current node's
data/shape.

Now also using node data passed in as props by xyflow instead of
subscribing to it via re-frame. Just needs to be converted to the clj
version of the node data, but animation speed was greatly improved by
doing this.